requirements.txt
Django >= 2.2, < 2.2.2.0
djangorestframework >= 3.9.2, < 3.10.0
psycopg2 >= 2.8, < 2.9
celery
requests
django_cleanup
